+++ test/utils.js
@@ -1,9 +1,14 @@
 import _ from 'underscore';
 import test from 'tape';
 
-var Common = require('./common');
-var Elsinore = Common.Elsinore;
-var Utils = Elsinore.Utils;
+import {
+    Elsinore,
+    Utils,
+} from './common';
+// var Common = require('./common');
+// var Elsinore = Common.Elsinore;
+// var Utils = Elsinore.Utils;
+
 
 
 test('toPascalCase', function(t){
@@ -16,9 +21,17 @@ test('toPascalCase', function(t){
         'goodJobEveryone': 'GoodJobEveryone'
     };
 
-    _.each( cases, function(expected,input){
-        t.equals( Utils.toPascalCase(input), expected );
-    });
-
+    _.each( cases, (expected,input) => t.equals( Utils.toPascalCase(input), expected ));
     t.end();
 });
+
+
+test('parse uri', t => {
+    const record = Utils.parseUri('/foo/bar?baz=quux#frag');
+
+    t.deepEqual( record.query, { 'baz': 'quux'} );
+    t.equals( record.baseUri, '/foo/bar?baz=quux' );
+    t.equals( record.fragment, 'frag' );
+    
+    t.end();
+})